Refactor toDoList web app to include Mongodb and render db items
Added Mongodb to store the item data sent from toDoList web app.

When the main route is loaded up for the first time, create the default items in Mongodb and redirect user back to home page to display the default items.

Otherwise, render the home page using EJS with the current date and all the data stored for the Item collection.